#38 restructure file layout for ducks pattern
